Rename crypto module to hasher
