+++ src/hasher.rs
